// api_services/categories/categories.interface.ts
// Generated by https://quicktype.io

export interface CategoresDto {
  id: number;
  title: string;
  image_location: string;
  created_at: string;
  children: any[];
}

// Generated by https://quicktype.io

export interface MainBannersDto {
  "section-1": Empty[];
  "section-2": Empty[];
}

export interface Empty {
  id: number;
  category_id: null;
  image_location: string;
  title: string;
  group: string;
  description: string;
  position: null | string;
  link: null;
  expire_date: null;
  created_at: string;
}

export interface CategoriesProdctBodyDto {
  storeId?: string;
  page?: string | number;
  tag_ids?: string[] | number[] | null;
  brand_id?: string | number | null;
  max_price?: string | number | null;
  min_price?: string | number | null;
  q?: string | null;
  sort_by?: string | null;
  min_unit_value?: string | number | null;
  max_unit_value?: string | number | null;
  unit_id?: string | number | null;
  category_id?: string | number | null | string[];
}

// Generated by https://quicktype.io
export interface MetaType {
  current_page: number;
  from: number;
  last_page: number;
  links: Link[];
  path: string;
  per_page: number;
  to: number;
  total: number;
}
export interface Link {
  url: null | string;
  label: string;
  active: boolean;
}

export interface CategoriesProductResponseDto {
  data: CategoriesProductResponseDataDto[];
  meta: MetaType;
}
export interface CategoriesProductResponseDataDto {
  id: number;
  unit_value: number;
  buy_price: number;
  sell_price: number;
  discount_amount: number;
  discount_percentage: string;
  view_count: number;
  is_special: boolean;
  is_new: boolean;
  is_active: boolean;
  is_favorite: boolean;
  in_cart: number;
  unit: Unit;
  product: Product;
  business: Business;
}

export interface Business {
  id: number;
  title: string;
  address: string;
  rate: string;
  rate_count: number;
  location: Location;
  post_code: string;
  email: string;
  logo_location: string;
  has_deposit: boolean;
  send_price: number;
  send_type: string;
  coverage_post_codes: string[];
  commission_percentage: string;
  username: string;
  created_at: string;
}

export interface Location {
  latitude: string;
  longitude: string;
}

export interface Product {
  id: number;
  title: string;
  description: string;
  cover_location: CoverLocation;
  unit_value: number;
  code: string;
  barcode: string;
  rate: string;
  rate_count: number;
  buy_price: number;
  sell_price: number;
  discount_amount: number;
  discount_percentage: string;
  media_files: CoverLocation[];
  depositItem: DepositItem;
}

export interface CoverLocation {
  id: string;
  sort_number: null;
  file_location: string;
  file_mime_type: string;
}

export interface DepositItem {
  id: number;
  title: string;
  deposit_price: number;
  image_location: string;
}

export interface Unit {
  id: number;
  title: string;
  key: string;
}

// Generated by https://quicktype.io

export interface BrandsDto {
  id: number;
  title: string;
  image_location: string;
}